requirements.txt
docling
openai
python-dotenv
pydantic
requests
tqdm
typing_extensions>=4.12
pydantic>=2.7
sentence-transformers
scikit-learn
umap-learn
hdbscan
